Refactor Localeapp::Routes (split it up in multiple files, actually)
